Register, login and logout working
