+++ apps/server/auth/src/main/java/com/archeo/auth/controllers/AuthController.java
@@ -1,0 +1,43 @@
+package com.archeo.auth.controllers;
+
+
+import com.archeo.auth.dtos.AuthResponse;
+import com.archeo.auth.dtos.SigninRequest;
+import com.archeo.auth.dtos.SignupRequest;
+import com.archeo.auth.services.AuthService;
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/api/auth")
+@RequiredArgsConstructor
+public class AuthController {
+
+
+    private final AuthService authService;
+
+    @PostMapping("/register")
+    public ResponseEntity<AuthResponse> register(@RequestBody SignupRequest signupRequest, HttpServletRequest request){
+        AuthResponse response=authService.register(signupRequest, request);
+        return ResponseEntity.ok(response);
+    }
+
+    @PostMapping("/login")
+    public ResponseEntity<AuthResponse> login(@RequestBody SigninRequest signinRequest, HttpServletRequest request){
+        AuthResponse response=authService.login(signinRequest, request);
+        return ResponseEntity.ok(response);
+    }
+
+    @PostMapping("/logout")
+    public ResponseEntity<String> logout(@RequestHeader("Authorization") String authHeader){
+        String token=authHeader.replace("Bearer ", "");
+        String response=authService.logout(token);
+        return ResponseEntity.ok(response);
+
+    }
+
+
+
+}
