+++ apps/server/auth/src/main/java/com/archeo/auth/dtos/SigninRequest.java
@@ -1,0 +1,12 @@
+package com.archeo.auth.dtos;
+
+import lombok.Data;
+
+@Data
+
+public class SigninRequest {
+
+    private String email;
+    private String username;
+    private String password;
+}
