+++ apps/server/auth/src/main/java/com/archeo/auth/dtos/SignupRequest.java
@@ -1,0 +1,11 @@
+package com.archeo.auth.dtos;
+
+import lombok.Data;
+
+@Data
+public class SignupRequest {
+
+    private String username;
+    private String email;
+    private String password;
+}
