+++ apps/server/auth/src/main/java/com/archeo/auth/repositories/SessionRepo.java
@@ -1,0 +1,17 @@
+package com.archeo.auth.repositories;
+
+import com.archeo.auth.models.Session;
+import com.archeo.user.models.Users;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.UUID;
+
+@Repository
+public interface SessionRepo extends JpaRepository<Session, UUID> {
+
+    List<Session> findByUser(Users user);
+    void deleteByRefreshTokenHash(String refreshTokenHash);
+
+}
