+++ apps/server/auth/src/main/java/com/archeo/auth/services/AuthLogsService.java
@@ -1,0 +1,29 @@
+package com.archeo.auth.services;
+
+import com.archeo.auth.repositories.AuthLogsRepo;
+import com.archeo.auth.models.AuthLogs;
+import com.archeo.user.models.Users;
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class AuthLogsService {
+
+    private final AuthLogsRepo authLogsRepo;
+    private final PasswordEncoder passwordEncoder;
+
+    public void log(Users user, String refreshToken, HttpServletRequest request){
+
+        AuthLogs authLogs=AuthLogs.builder()
+                .user(user)
+                .refreshTokenHash(passwordEncoder.encode(refreshToken))
+                .ipAddress(request.getRemoteAddr())
+                .userAgent(request.getHeader("User-Agent"))
+                .build();
+
+        authLogsRepo.save(authLogs);
+    }
+}
