--- apps/server/auth/src/main/java/com/archeo/server/modules/auth/config/JwtConstant.java
@@ -1,11 +1,0 @@
-package com.archeo.server.modules.auth.config;
-
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class JwtConstant {
-
-    public static final String SECRET_KEY = "jbfvbjbhjbfjkdbjkiebfibvifbvibfibiifbvhi";
-
-    public static final String AUTHORIZATION_HEADER = "Authorization";
-}
