--- apps/server/auth/src/main/java/com/archeo/server/modules/auth/config/JwtTokenValidator.java
@@ -1,78 +1,0 @@
-package com.archeo.server.modules.auth.config;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.AuthorityUtils;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import javax.crypto.SecretKey;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.util.List;
-
-@Configuration
-public class JwtTokenValidator extends OncePerRequestFilter {
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
-            throws ServletException, IOException {
-
-        String requestURI = request.getRequestURI();
-
-        // Skip JWT validation for public endpoints
-        if (requestURI.startsWith("/api/auth/signIn") || requestURI.startsWith("/api/auth/signUp")) {
-            filterChain.doFilter(request, response);
-            return;
-        }
-
-        String jwt = request.getHeader("Authorization");
-
-        if (jwt != null && jwt.startsWith("Bearer ")) {
-            jwt = jwt.substring(7); // Remove "Bearer " prefix
-
-            try {
-                SecretKey key = Keys.hmacShaKeyFor(JwtConstant.SECRET_KEY.getBytes(StandardCharsets.UTF_8));
-
-                Claims claims = Jwts.parserBuilder()
-                        .setSigningKey(key)
-                        .build()
-                        .parseClaimsJws(jwt)
-                        .getBody();
-
-                String email = claims.getSubject(); // Assumes 'sub' claim holds the email
-                String authorities = String.valueOf(claims.get("authorities"));
-
-                System.out.println("Authorities: "+authorities);
-
-                List<GrantedAuthority> authorityList = AuthorityUtils.commaSeparatedStringToAuthorityList(authorities);
-
-                System.out.println("Authorities from token: " + authorityList);  // Log authorities to check correctness
-
-                Authentication auth = new UsernamePasswordAuthenticationToken(
-                        email,
-                        null,
-                        authorityList
-                );
-
-                SecurityContextHolder.getContext().setAuthentication(auth);
-            } catch (Exception e) {
-                response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-                response.getWriter().write("Invalid or expired token");
-                return;
-            }
-        }
-
-        filterChain.doFilter(request, response);
-    }
-
-}
