--- apps/server/auth/src/main/java/com/archeo/server/modules/auth/dtos/AuthResponse.java
@@ -1,18 +1,0 @@
-package com.archeo.server.modules.auth.dtos;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class AuthResponse {
-
-    private String accessToken;
-    private String refreshToken;
-    private String userRole;
-
-}
