--- apps/server/auth/src/main/java/com/archeo/server/modules/auth/dtos/SigninRequest.java
@@ -1,12 +1,0 @@
-package com.archeo.server.modules.auth.dtos;
-
-import lombok.Data;
-
-@Data
-
-public class SigninRequest {
-
-    private String email;
-    private String username;
-    private String password;
-}
