--- apps/server/auth/src/main/java/com/archeo/server/modules/auth/dtos/SignupRequest.java
@@ -1,11 +1,0 @@
-package com.archeo.server.modules.auth.dtos;
-
-import lombok.Data;
-
-@Data
-public class SignupRequest {
-
-    private String username;
-    private String email;
-    private String password;
-}
