--- apps/server/auth/src/main/java/com/archeo/server/modules/auth/models/Session.java
@@ -1,39 +1,0 @@
-package com.archeo.server.modules.auth.models;
-
-import com.archeo.server.modules.user.models.Users;
-import jakarta.persistence.*;
-import lombok.Data;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.sql.Timestamp;
-import java.util.UUID;
-
-@Entity
-@Table(name = "sessions")
-@Data
-public class Session {
-
-    @Id
-    @GeneratedValue
-    private UUID id;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "user_id", referencedColumnName = "id", nullable = false)
-    private Users user;
-
-    @Column(name = "refresh_token", nullable = false, length = 255)
-    private String refreshTokenHash;
-
-    @Column(name = "ip_address", length = 45)
-    private String ipAddress;
-
-    @Column(name = "user_agent", length = 255)
-    private String userAgent;
-
-    @Column(name = "expires_at", nullable = false)
-    private Timestamp expiresAt;
-
-    @CreationTimestamp
-    @Column(name = "created_at", updatable = false)
-    private Timestamp createdAt;
-}
