--- apps/server/user/src/main/java/com/archeo/server/modules/user/config/SecurityConfig.java
@@ -1,8 +1,0 @@
-package com.archeo.server.modules.user.config;
-
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SecurityConfig{
-
-}
