+++ apps/server/user/src/main/java/com/archeo/user/config/SecurityConfig.java
@@ -1,0 +1,8 @@
+package com.archeo.user.config;
+
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SecurityConfig{
+
+}
