+++ apps/server/user/src/main/java/com/archeo/user/repositories/UserRepository.java
@@ -1,0 +1,14 @@
+package com.archeo.user.repositories;
+
+import com.archeo.user.models.Users;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<Users, Long> {
+
+    Optional<Users> findByEmail(String email);
+    Optional<Users> findByUsername(String username);
+}
